extension support for chrome browser
the chrome driver wrapper will pass a list of extensions to the driver that are supposed to be loaded during initialization.
Equal to the handling for firefox in splinter/splinter/driver/webdriver/firefox.py the chrome WebDriver will now have an additional parameter that can contain a list of extensions, that are to be loaded. This parameter is None for default and will only be taken into account, if it is present.