Enable timeout setting for Firefox webdriver
Currently there's no straightforward way to specify the timeout for the Firefox webdriver. 

Selenium's default value (30) is too short for low-performance settings, so it would be desirable to not only have a larger timeout window (this PR sets it to 90) as well as letting the user define this when instantiating the webdriver.