Added retry_count to get_driver (#754)
* Added retry_count to get_driver

It looks like Browser is taking `retry_count` as an argument
which is never used. Fixed by adding it to get_driver.

* Added test to make sure retry_count is being used

Checks that retry_count is being used in Browser by calling a test driver
that will raise an error with the number of times it has run. The final
error returned should then contain the retry_count.

* fixed typo