Update index.rst
name='btnG' does not exist on "http://www.google.com" and a run of the sample code results in an IndexError exception.  I propose that the button be accessed by value='Google Search' as this currently works and will probably be more resilient to change than the Google element names.