Modified sample code in documentation with one in index file in Github
The Sample code in documentation wasn't updated as the button name in Google search was modified to `btnK` from `btnG` and the code was not running well due to that. But the modified sample code was in Github `index.rst`. So the same sample code is used in documentation too as it will be helpful for new people.