github/workflows: fix selenium-server flakiness
The previous code wasn't doing the right thing: it started the server in
background, slept for 3 seconds in background and started the tests.
Adding another `&` after `sleep 3` would make it actually wait 3
seconds, but instead of doing that we use `timeout` and try the
healthcheck endpoint until the server is ready.

Also upgrade selenium-server.